Implement Semantic Similarity Detection Using SBERT
