+++ app/schemas/issue_schema.py
@@ -1,0 +1,11 @@
+from dataclasses import dataclass
+from typing import List, Optional
+
+@dataclass
+class Issue:
+    number: int
+    title: str
+    url: str
+    state: str
+    comments: List[str]
+    threshold: Optional[float] = None
