+++ requirements.txt
@@ -7,7 +7,10 @@ colorama==0.4.6
 coverage==7.6.4
 dill==0.3.9
 exceptiongroup==1.2.2
+filelock==3.16.1
 Flask==3.0.3
+fsspec==2024.10.0
+huggingface-hub==0.26.2
 idna==3.10
 iniconfig==2.0.0
 isort==5.13.2
@@ -16,20 +19,32 @@ Jinja2==3.1.4
 joblib==1.4.2
 MarkupSafe==3.0.2
 mccabe==0.7.0
+mpmath==1.3.0
+networkx==3.4.2
 numpy==2.1.2
 packaging==24.1
+pillow==11.0.0
 platformdirs==4.3.6
 pluggy==1.5.0
 pylint==3.3.1
 pytest==8.3.3
 pytest-cov==6.0.0
 python-dotenv==1.0.1
+PyYAML==6.0.2
+regex==2024.9.11
 requests==2.32.3
+safetensors==0.4.5
 scikit-learn==1.5.2
 scipy==1.14.1
+sentence-transformers==3.2.1
+sympy==1.13.1
 threadpoolctl==3.5.0
+tokenizers==0.20.1
 tomli==2.0.2
 tomlkit==0.13.2
+torch==2.5.1
+tqdm==4.66.6
+transformers==4.46.1
 typing_extensions==4.12.2
 urllib3==2.2.3
-Werkzeug==3.0.6
+Werkzeug==3.1.1
