+++ run.py
@@ -4,5 +4,5 @@ app = create_app()
 
 if __name__ == '__main__':
     host = app.config.get('HOST', '127.0.0.1')
-    port = int(app.config.get('PORT', 5000))
+    port = int(app.config.get('PORT') or 5000)
     app.run(host=host, port=port, debug=True)
